requirements.txt
numpy==1.23.5
Pillow==9.3.0
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
tqdm==4.64.1